feat(router): enable OCSP stapling by default
This allows nginx to cache the [OCSP] response from the Certificate Authority
and return a signed copy of it to the User Agent. This reduces the
number of roundtrips necessary to verify the certificate and therefore
results in both a performance improvement and reduced load on the OCSP
server.

If an invalid response is stapled and sent to the client it is ignored
so disabling verification by default does no harm.

This currently requires `/deis/router/ocsp/responder` to be manually set to the
IP Address of the certificate's OCSP server. This will not be necessary if
a dns resolver is configured so nginx can look up the domain name specified
by the certificate.

[OCSP]: https://tools.ietf.org/html/rfc2560